requirements.txt
anyio==3.5.0
asgiref==3.5.0
atomicwrites==1.4.0
attrs==21.4.0
bcrypt==3.2.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.2
colorama==0.4.4
coverage==6.3.2
cryptography==37.0.1
dependency-injector==4.39.1
ecdsa==0.17.0
fastapi==0.75.2
fastapi-events==0.3.0
fastapi-jwt==0.1.8
fastapi-route-log==0.0.7
fastapi-utils==0.2.1
greenlet==1.1.2
h11==0.13.0
httptools==0.4.0
idna==3.3
iniconfig==1.1.1
packaging==21.3
passlib==1.7.4
pluggy==1.0.0
psycopg2==2.9.3
py==1.11.0
pyasn1==0.4.8
pycparser==2.21
pydantic==1.9.0
pyparsing==3.0.8
pytest==7.1.2
pytest-cov==3.0.0
python-dotenv==0.20.0
python-jose==3.3.0
python-multipart==0.0.5
PyYAML==6.0
requests==2.27.1
rsa==4.8
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.35
starlette==0.17.1
tomli==2.0.1
typing_extensions==4.2.0
ua-parser==0.10.0
urllib3==1.26.9
user-agents==2.2.0
uvicorn==0.17.6
watchgod==0.8.2
websockets==10.3
